+++ CHANGELOG.md
@@ -3,3 +3,21 @@
 ## develop branch
 
 Nothing yet.
+
+## 1.0.0 - Thu Sep 10 2015
+
+### New
+
+* Checks\HasEventHandler
+* Checks\IsEventName
+* Events\Event
+* Exceptions\Exxx_EventStreamException
+* Exceptions\E4xx_EventStreamException
+* Exceptions\E4xx_NotAnEvent
+* Exceptions\E4xx_UnsupportedType
+* Requirements\RequireEventName
+* Streams\DispatchEvent
+* Streams\EventStream
+* Streams\RegisterEventHandler
+* ValueBuilders\GetEventHandlerList
+* ValueBuilders\GuaranteeEventStream